Add files via upload
🔐 *Project Title:*

*Dynamic Two-Factor Authentication (2FA) Based on Contextual Risk Analysis*

---

## 📄 *Summary:*

Most current 2FA systems apply the same level of authentication regardless of the context of the login attempt. This project introduces a *dynamic 2FA system* that evaluates the *risk level of each login* based on factors such as *IP address, **time of access, and **device. Based on the risk score, the system adjusts the second layer of authentication — **from no 2FA, to **OTP, to **biometric/manual approval*.